polish token logo
