+++ src/components/Convertor/Pool.tsx
@@ -1,16 +1,7 @@
-import {
-  Button,
-  Flex,
-  Image,
-  Link,
-  Text,
-  useColorModeValue,
-} from '@chakra-ui/react'
+import { Button, Flex, useColorModeValue } from '@chakra-ui/react'
 import { ConversionPool, FungibleTokenMetadata } from 'types'
-import { DecimalUtil } from 'utils'
-import { MdSyncAlt, MdTrendingFlat } from 'react-icons/md'
 import { isMobile } from 'react-device-detect'
-import { useGlobalStore } from 'stores'
+import PoolInfo from './PoolInfo'
 
 export default function Pool({
   pool,
@@ -22,23 +13,6 @@ export default function Pool({
   onSelect: (pool: ConversionPool) => void
 }) {
   const bg = useColorModeValue('white', '#25263c')
-  const inToken = whitelist.find((t) => t.token_id === pool.in_token)
-  const outToken = whitelist.find((t) => t.token_id === pool.out_token)
-
-  const { global } = useGlobalStore()
-  const inTokenLiq = DecimalUtil.fromString(
-    pool.in_token_balance,
-    inToken?.decimals
-  )
-    .toFixed(2)
-    .toString()
-
-  const outTokenLiq = DecimalUtil.fromString(
-    pool.out_token_balance,
-    outToken?.decimals
-  )
-    .toFixed(2)
-    .toString()
 
   return (
     <Flex
@@ -50,58 +24,7 @@ export default function Pool({
       justify="space-between"
       mb={2}
     >
-      <Flex direction="column" gap={2} flexShrink={0}>
-        <Text color="#008cd5">
-          {`#${pool.id} Owner: `}
-          <Link
-            href={`${global.network?.near.explorerUrl}/accounts/${pool.creator}`}
-          >
-            {pool.creator}
-          </Link>
-        </Text>
-        <Flex direction="row" align="flex-start" gap={4}>
-          <Flex direction="column" gap={2}>
-            <Flex gap={2} align="center">
-              {inToken && (
-                <Image src={inToken.icon || ''} width={10} height={10} alt="" />
-              )}
-              <Text fontSize="2xl">{inToken?.symbol}</Text>
-            </Flex>
-            {pool.reversible && (
-              <Text
-                fontSize="sm"
-                className="octo-gray"
-              >{`Liquidity: ${inTokenLiq}`}</Text>
-            )}
-          </Flex>
-          <Flex direction="column" align="center">
-            <Text fontSize="sm">{`${pool.in_token_rate} : ${pool.out_token_rate}`}</Text>
-            {pool.reversible ? (
-              <MdSyncAlt size={30} className="octo-gray" />
-            ) : (
-              <MdTrendingFlat size={30} className="octo-gray" />
-            )}
-          </Flex>
-          <Flex direction="column" gap={2}>
-            <Flex gap={2} align="center">
-              {outToken && (
-                <Image
-                  src={outToken.icon || ''}
-                  width={10}
-                  height={10}
-                  alt=""
-                />
-              )}
-              <Text fontSize="2xl">{outToken?.symbol}</Text>
-            </Flex>
-
-            <Text
-              fontSize="sm"
-              className="octo-gray"
-            >{`Liquidity: ${outTokenLiq}`}</Text>
-          </Flex>
-        </Flex>
-      </Flex>
+      <PoolInfo pool={pool} whitelist={whitelist} />
       <Button
         variant="octo-linear"
         alignSelf="flex-end"
